bot_cripto/requirements.txt
python-telegram-bot
requests
python-dotenv
telebot
pandas
schedule
data-fetcher
pyTelegramBotAPI
matplotlib 
numpy
yfinance